feat: sketch of v2.
Reworked server to run either V1 or V2. Simplified health end point.
Added a metrics end point, but no metrics as of yet. Attempting to use a
an upstream library for the GRPC code. Not running, but compling. Want
to check it in because I will probably have to delete things to get it
to run.